rename package to foo
