+++ app/upgrades/foo/upgrade.go
@@ -1,4 +1,4 @@
-package drs5
+package foo
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
